requirements.txt
requests>=2.32.0
praw>=7.8.0
python-dotenv>=1.0.1
tzdata>=2025.1
